Updates PWA configuration for correct routing
Ensures the PWA functions correctly when deployed to a subdirectory.
This change modifies the manifest and service worker to correctly
handle routing within the PocketCaddy application when it's not
hosted at the root of the domain. Specifically, it updates the
`start_url` and `scope` in `manifest.json` and adjusts the `BASE_PATH`
in `service-worker.js`.  Also adds a handler to fetch specifically Blazor framework files in the service worker.