Improves iOS app snapshot on background
Ensures iOS captures a fresh app snapshot when the app is backgrounded.

This prevents the display of stale content in the app switcher and improves
the user experience when switching back to the app. It achieves this by
injecting a dummy element and then removing it after a short delay, triggering
a snapshot refresh.

Also updates the base href and some styling to improve the app's
appearance.