Adds experimental side offset calculation
Adds an option to include an experimental side offset calculation for rollout values.
This feature estimates the lateral movement of the ball due to wind and pull,
expressed as the number of cup widths to the side of the hole's center.

This is a work in progress and currently calibrated for the Eclipse 10 scope.

Also improves table layouts by using css flexbox, making table column
sizes uniform.